+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased Changes
 
+### Changed
+
+- The PixelInspector now shows the lat/lon of the clicked location.
+
 ## v0.9.6 - 11/08/23
 
 ### Changed